Remove mixin and add header logic to middleware
