+++ readthedocs/proxito/views/mixins.py
@@ -22,43 +22,6 @@ from readthedocs.redirects.exceptions import InfiniteRedirectException
 log = logging.getLogger(__name__)  # noqa
 
 
-class ProxitoHeaderMixin:
-
-    def add_proxito_headers(self, request, response):
-        """Add debugging headers to proxito responses."""
-
-        # Pull data off the request and response
-        project_slug = getattr(request, 'host_project_slug', '')
-        version_slug = getattr(request, 'path_version_slug', '')
-        path = getattr(response, 'proxito_path', '')
-
-        response['X-RTD-Domain'] = request.get_host()
-        response['X-RTD-Project'] = project_slug
-
-        if version_slug:
-            response['X-RTD-Version'] = version_slug
-
-        # Some responses don't have a path like redirects
-        if path:
-            response['X-RTD-Path'] = path
-
-        # Include the project & project-version so we can do larger purges if needed
-        response['Cache-Tag'] = f'{project_slug}'
-        if version_slug:
-            response['Cache-Tag'] += f',{project_slug}-{version_slug}'
-        # Larger debug headers
-        if hasattr(request, 'rtdheader'):
-            response['X-RTD-Project-Method'] = 'rtdheader'
-        elif hasattr(request, 'subdomain'):
-            response['X-RTD-Project-Method'] = 'subdomain'
-        elif hasattr(request, 'cname'):
-            response['X-RTD-Project-Method'] = 'cname'
-        if hasattr(request, 'external_domain'):
-            response['X-RTD-Version-Method'] = 'domain'
-        else:
-            response['X-RTD-Version-Method'] = 'path'
-
-
 class ServeDocsMixin:
 
     """Class implementing all the logic to serve a document."""
